put multiple used functions in new module
